refactored/requirements.txt
# 核心依赖
# pygame>=2.0.0
pygame-gui

# 可选依赖（视频背景支持）
opencv-python

# 开发依赖
pytest
black
flake8

# 文档生成
sphinx
